Reorganize utils
- Moved utils.ts to utils/index.ts
- Moved timeUtils.ts to utils/time.ts
- Moved sorting functions to utils/sorting.ts
