+++ source/SIL.AppBuilder.Portal/src/lib/products/components/BuildArtifacts.svelte
@@ -3,8 +3,9 @@
   import Tooltip from '$lib/components/Tooltip.svelte';
   import * as m from '$lib/paraglide/messages';
   import { languageTag } from '$lib/paraglide/runtime';
-  import { getRelativeTime } from '$lib/timeUtils';
-  import { byString, bytesToHumanSize } from '$lib/utils';
+  import { bytesToHumanSize } from '$lib/utils';
+  import { byString } from '$lib/utils/sorting';
+  import { getRelativeTime } from '$lib/utils/time';
 
   interface Props {
     build: {
