+++ source/SIL.AppBuilder.Portal/src/lib/projects/common.ts
@@ -1,5 +1,5 @@
 import { paginateSchema } from '$lib/table';
-import { isAdminForOrg, isSuperAdmin } from '$lib/utils';
+import { isAdminForOrg, isSuperAdmin } from '$lib/utils/roles';
 import { idSchema } from '$lib/valibot';
 import type { Session } from '@auth/sveltekit';
 import type { Prisma } from '@prisma/client';
