+++ source/SIL.AppBuilder.Portal/src/lib/projects/components/ProjectCard.svelte
@@ -3,8 +3,8 @@
   import { getIcon } from '$lib/icons/productDefinitionIcon';
   import * as m from '$lib/paraglide/messages';
   import { languageTag } from '$lib/paraglide/runtime';
-  import { getTimeDateString } from '$lib/timeUtils';
-  import { byString } from '$lib/utils';
+  import { byString } from '$lib/utils/sorting';
+  import { getTimeDateString } from '$lib/utils/time';
   import type { PrunedProject } from '../common';
 
   interface Props {
