+++ source/SIL.AppBuilder.Portal/src/routes/(authenticated)/users/[id=idNumber]/settings/roles/+page.server.ts
@@ -1,4 +1,4 @@
-import { isAdminForOrg, isSuperAdmin, orgsForRole } from '$lib/utils';
+import { isAdminForOrg, isSuperAdmin, orgsForRole } from '$lib/utils/roles';
 import { idSchema } from '$lib/valibot';
 import { error } from '@sveltejs/kit';
 import { DatabaseWrites, prisma } from 'sil.appbuilder.portal.common';
