+++ source/SIL.AppBuilder.Portal/src/routes/(authenticated)/users/invite/+page.server.ts
@@ -1,4 +1,4 @@
-import { isAdminForOrg, isSuperAdmin } from '$lib/utils';
+import { isAdminForOrg, isSuperAdmin } from '$lib/utils/roles';
 import { idSchema } from '$lib/valibot';
 import { fail } from '@sveltejs/kit';
 import { DatabaseWrites, prisma } from 'sil.appbuilder.portal.common';
