+++ source/SIL.AppBuilder.Portal/src/routes/(authenticated)/workflow-instances/+page.svelte
@@ -7,8 +7,8 @@
   import Tooltip from '$lib/components/Tooltip.svelte';
   import * as m from '$lib/paraglide/messages';
   import { languageTag } from '$lib/paraglide/runtime';
-  import { getRelativeTime } from '$lib/timeUtils';
-  import { byName } from '$lib/utils';
+  import { byName } from '$lib/utils/sorting';
+  import { getRelativeTime } from '$lib/utils/time';
   import type { FormResult } from 'sveltekit-superforms';
   import { superForm } from 'sveltekit-superforms';
   import type { PageData } from './$types';
